code/requirements.txt
pygame==2.6.1
pyserial==3.5